requirements_minimal.txt
# RAG Python v1.5.1 - Requisitos Mínimos para Inicialização
# Dependências essenciais para o inicializador robusto funcionar

# === ESSENCIAIS PARA INICIALIZAÇÃO ===
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0
psutil>=5.9.0

# === CORE SYSTEM ===
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.6.0

# === DATABASES ===
psycopg2-binary>=2.9.7
chromadb>=0.4.15

# === DOCUMENT PROCESSING ===
pypdf2>=3.0.1
python-docx>=0.8.11
pandas>=2.1.0

# === PRIVACY ===
presidio-analyzer>=2.2.33
presidio-anonymizer>=2.2.33

# === UTILITIES ===
tqdm>=4.66.0
colorama>=0.4.6 